publicProfile pages added
