+++ src/constants/urls.ts
@@ -12,7 +12,8 @@ export const constants = {
     GET_VACANCIES: 'search_work/vacancies/',
     GET_CATEGORIES: 'search_work/categories/',
     GET_SKILLS: 'search_work/skills/',
-    GET_PROFILE: '/profile',
+    GET_PROFILE: '/profile/',
+    PROFILE: '/profile/:id',
     GET_ALL_PROFILES: '/profile/all',
     POST_PROFILE: '/profile/update',
     USER_CONTACTS: '/user/contacts',
